+++ src/habr_tech_radar/filtering/__init__.py
@@ -1,3 +1,4 @@
+from habr_tech_radar.filtering.heuristic import HeuristicArticleFilter
 from habr_tech_radar.filtering.service import ArticleFilter, StubArticleFilter
 
-__all__ = ["ArticleFilter", "StubArticleFilter"]
+__all__ = ["ArticleFilter", "HeuristicArticleFilter", "StubArticleFilter"]